Task5 done

+++ src/main/java/com/stackroute/MovieServiceApplication/service/MovieServiceImpl.java
@@ -1,5 +1,7 @@
 package com.stackroute.MovieServiceApplication.service;
 
+import com.stackroute.MovieServiceApplication.customException.MovieAlreadyExistException;
+import com.stackroute.MovieServiceApplication.customException.MovieNotFoundException;
 import com.stackroute.MovieServiceApplication.domain.Movie;
 import com.stackroute.MovieServiceApplication.repository.MovieRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,28 +28,32 @@ public class MovieServiceImpl implements MovieService {
     }
 
     @Override
-    public Movie getMovieById(int movieId) {
+    public Movie getMovieById(int movieId) throws MovieNotFoundException{
+        if(!movieRepository.existsById(movieId)) throw new MovieNotFoundException("Movie with ID: "+movieId+" doesn't Exist.");
         return movieRepository.findById(movieId).get();
     }
 
     @Override
-    public List<Movie> getMovieByTitle(String movieTitle) {
+    public List<Movie> getMovieByTitle(String movieTitle){
         return movieRepository.findByTitle(movieTitle);
     }
 
     @Override
-    public List<Movie> deleteMovie(int movieId) {
+    public List<Movie> deleteMovie(int movieId) throws MovieNotFoundException {
+        if(!movieRepository.existsById(movieId)) throw new MovieNotFoundException("Movie with ID: "+movieId+" doesn't Exist.");
         movieRepository.deleteById(movieId);
         return getAllMovie();
     }
 
     @Override
-    public Movie saveMovie(Movie movie) {
+    public Movie saveMovie(Movie movie) throws MovieAlreadyExistException {
+        if(!movieRepository.findByTitleAndDate(movie.getMovieTitle(),movie.getMovieReleaseDate()).isEmpty()) throw new MovieAlreadyExistException("Movie with ID: "+movie.getMovieId()+" Already Exist.");
         return movieRepository.save(movie);
     }
 
     @Override
-    public Movie updateMovie(Movie movie) {
+    public Movie updateMovie(Movie movie) throws MovieNotFoundException {
+        if(!movieRepository.existsById(movie.getMovieId())) throw new MovieNotFoundException("Movie with ID: "+movie.getMovieId()+" doesn't Exist.");
         Movie movie1 = movieRepository.getOne(movie.getMovieId());
         if(movie.getMovieTitle()!=null) movie1.setMovieTitle(movie.getMovieTitle());
         if(movie.getMovieOverview()!= null) movie1.setMovieOverview(movie.getMovieOverview());
